Split out shouldFilterWpcomThemes
