+++ client/state/themes/selectors/index.js
@@ -7,12 +7,7 @@ import i18n from 'i18n-calypso';
 /**
  * Internal dependencies
  */
-import {
-	getSiteSlug,
-	isJetpackSite,
-	hasJetpackSiteJetpackThemesExtendedFeatures,
-	isJetpackSiteMultiSite,
-} from 'state/sites/selectors';
+import { getSiteSlug, isJetpackSite } from 'state/sites/selectors';
 import { hasFeature } from 'state/sites/plans/selectors';
 import { getThemeTaxonomySlugs } from 'state/themes/utils';
 import { FEATURE_UNLIMITED_PREMIUM_THEMES } from 'lib/plans/constants';
@@ -58,24 +53,7 @@ export { getThemePreviewThemeOptions } from 'state/themes/selectors/get-theme-pr
 export { themePreviewVisibility } from 'state/themes/selectors/theme-preview-visibility';
 export { getWpcomParentThemeId } from 'state/themes/selectors/get-wpcom-parent-theme-id';
 export { isDownloadableFromWpcom } from 'state/themes/selectors/is-downloadable-from-wpcom';
-
-/**
- * Determine whether wpcom themes should be removed from
- * a queried list of themes. For jetpack sites with the
- * required capabilities, we hide wpcom themes from the
- * list of locally installed themes.
- *
- * @param {object} state Global state tree
- * @param {number} siteId The Site ID
- * @returns {boolean} true if wpcom themes should be removed
- */
-export function shouldFilterWpcomThemes( state, siteId ) {
-	return (
-		isJetpackSite( state, siteId ) &&
-		hasJetpackSiteJetpackThemesExtendedFeatures( state, siteId ) &&
-		! isJetpackSiteMultiSite( state, siteId )
-	);
-}
+export { shouldFilterWpcomThemes } from 'state/themes/selectors/should-filter-wpcom-themes';
 
 /**
  * Returns the URL for purchasing a Jetpack Professional plan if the theme is a premium theme and site doesn't have access to them.
